walsvid 2to3

+++ lib/fast_rcnn/test.py
@@ -155,7 +155,7 @@ def im_detect(sess, net, im, boxes=None):
     # on the unique subset.
     if cfg.DEDUP_BOXES > 0 and not cfg.TEST.HAS_RPN:
         v = np.array([1, 1e3, 1e6, 1e9, 1e12])
-        hashes = np.round(blobs['rois'] * cfg.DEDUP_BOXES).dot(v)
+        hashes = int(np.round(blobs['rois'] * cfg.DEDUP_BOXES).dot(v))
         _, index, inv_index = np.unique(hashes, return_index=True,
                                         return_inverse=True)
         blobs['rois'] = blobs['rois'][index, :]
@@ -339,9 +339,9 @@ def test_net(sess, net, imdb, weights_filename, max_per_image=300, thresh=0.05, 
                     all_boxes[j][i] = all_boxes[j][i][keep, :]
         _t['misc'].toc()
 
-        print('im_detect: {:d}/{:d} {:.3f}s {:.3f}s'
-              .format(i + 1, num_images, _t['im_detect'].average_time,
-                      _t['misc'].average_time))
+        print('im_detect: {:d}/{:d} {:.3f}s {:.3f}s'.format(i + 1, num_images,
+                                                            _t['im_detect'].average_time,
+                                                            _t['misc'].average_time))
 
     det_file = os.path.join(output_dir, 'detections.pkl')
     with open(det_file, 'wb') as f:
